pyproject.toml
[tool.poetry]
name = "afdd"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.19"
rdflib = "^7.0.0"
pytest = "^8.2.2"
pandas = "^2.2.2"


python-dotenv = "^1.0.1"
neo4j = "^5.22.0"
numpy = "1.26.4"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
afdd = "afdd.main:main"
sample_data = "sample_data.create_sample_data:main"